[DEV] h264_filter: macroblock status and video stats
Macroblock statuses indicate for each MB of the current frame if
the MB is within a valid I-slice or P-slice, in a missing slice,
a concealed missing slice or within an error propagation.
The macroblock status array for the current frame can only be
retrieved withing the call to the auReady callback function.

Video stats are cumulative and based on macroblock statuses.
They include frame counters (total, output, discarded, missed),
error second counters (global and zone-based), and zone-based
macroblock status counters. The zones are 5 horizontal divisions
of the frame.

Change-Id: I9b78839ab0810ed4aa9f6a01e6ef56c7e17472b0
Reviewed-on: https://gerrit.parrot.biz/45094
Reviewed-by: Aurelien Barre <aurelien.barre@parrot.com>