[DEV] stream_sender: allow null target packet size
If target packet size is null, do not use aggregation (STAP-A):
only single-NALU and fragmentation (FU-A) to meet the maxPacketSize
requirement.
If target packet size is not null, use single-NALU, aggregation
(STAP-A) up to targetPacketSize and fragmentation (FU-A) to meet
both the targetPacketSize and maxPacketSize requirements.
Also fix wrong target packet size in dynamic config functions.

Change-Id: Ia2be54b6371d0c0c875354542be1afed83d2a785
Reviewed-on: https://gerrit.parrot.biz/57532
Reviewed-by: Djavan Bertrand <djavan.bertrand@parrot.com>